try fixing db tests again

+++ test/server.test.js
@@ -45,7 +45,7 @@ describe('Server', () => {
         });
 
         afterEach(async () => {
-            await new Promise((resolve, reject) => fs.rm(TEST_DB_FILE, (err) => err ? reject(err) : resolve()));
+            if (fs.existsSync(TEST_DB_FILE)) await new Promise((resolve, reject) => fs.rm(TEST_DB_FILE, (err) => err ? reject(err) : resolve()));
         });
 
         it('Should not return a value when asyncRun called', async () => {
@@ -97,7 +97,7 @@ describe('Server', () => {
         });
 
         afterEach(async () => {
-            await new Promise((resolve, reject) => fs.rm(TEST_DB_FILE, (err) => err ? reject(err) : resolve()));
+            if (fs.existsSync(TEST_DB_FILE)) await new Promise((resolve, reject) => fs.rm(TEST_DB_FILE, (err) => err ? reject(err) : resolve()));
         });
 
         it('Should return 400 Invalid Input when no token is provided', (t, done) => {
